+++ assets/controllers/certificate-limit_controller.js
@@ -1,0 +1,19 @@
+import { Controller } from '@hotwired/stimulus';
+
+export default class extends Controller {
+    static targets = ["input", "button"];
+    static values = {
+        profileLimitDate: Number
+    }
+
+    connect() {
+
+
+        console.log( this.profileLimitDateValue);
+    }
+
+    updateDate() {
+        this.inputTarget.value = this.profileLimitDateValue;
+
+    }
+}